# Use proto models directly for AggregateBatch