--- a/Makefile
+++ b/Makefile
@@ -44,7 +44,7 @@ gen-proto: $(PROTOC_GEN_GO) $(PROTOC_GEN_GO_VTPROTO) $(PROTOC)
 	$(eval PROTOC_VT_STRUCTS := $(shell for s in $(STRUCTS); do echo --go-vtproto_opt=pool=./aggregationpb.$$s ;done))
 	$(PROTOC) -I . --go_out=$(PROTOC_OUT) --plugin protoc-gen-go="$(PROTOC_GEN_GO)" \
 	--go-vtproto_out=$(PROTOC_OUT) --plugin protoc-gen-go-vtproto="$(PROTOC_GEN_GO_VTPROTO)" \
-	--go-vtproto_opt=features=marshal+unmarshal+size+pool \
+	--go-vtproto_opt=features=marshal+unmarshal+size+pool+clone \
 	$(PROTOC_VT_STRUCTS) \
 	$(wildcard proto/*.proto)
 	$(MAKE) fmt