… aggregators/aggregator_test.go
… aggregators/converter.go
… aggregators/converter_test.go
… aggregators/hasher_test.go
… aggregators/merger_test.go
… go.sum